# swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Vote online API
  version: 0.1.0
paths:
  /status:
    get:
      description: Get vote systems status
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
  /election/{id}/candidates:
    get:
      parameters:
        - in: path
          name: id
          required: true
      description: Get candidates
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - maxVotes
                  - lists
                properties:
                  maxVotes:
                    type: number
                    description: Max votes per user
                  lists:
                    type: array
                    items:
                      $ref: "#/components/schemas/List"
  /getCode:
    post:
      description: Get Code
      requestBody:
        description: The user to generate code.
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - electionID
                - name
                - surname
                - PESEL
              properties:
                electionID:
                  type: number
                name:
                  type: string
                surname:
                  type: string
                PESEL:
                  type: number

      responses:
        200:
          description: Success
          content:
            text/plain:
              schema:
                type: string
                example: code
        400:
          description: Error
          content:
            text/plain:
              schema:
                type: string
                example: Data invalid
  /vote:
    post:
      description: Make a vote
      requestBody:
        description: Candidates to add votes.
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
                - lists
              properties:
                code:
                  type: string
                lists:
                  type: array
                  items:
                    $ref: "#/components/schemas/List"
      responses:
        200:
          description: Success
        400:
          description: Code invalid

servers:
  - url: http://localhost:5000
    description: Production server (uses live data)
  - url: http://localhost:4000
    description: Sandbox server (uses test data)

components:
  schemas:
    Status:
      type: object
      properties:
        electionsState:
          $ref: "#/components/schemas/ElectionsState"
          description: Voting status
        startDate:
          type: string
          format: date
          description: Start date
        endDate:
          type: string
          format: date
          description: End date
    ElectionsState:
      type: string
      enum:
        - "inEdition"
        - "registered"
        - "ongoing"
        - "finished"
        - "closed"

    List:
      type: object
      properties:
        id:
          type: number
          description: List id
        name:
          type: string
          description: List name
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/Candidate"
    Candidate:
      type: object
      properties:
        id:
          type: number
          description: Candidate id
        name:
          type: string
          description: Candidate name
        surname:
          type: string
          description: Candidate surname